feat(ui): チケットページとチケットカードのコンポーネントを分割
